icarust integration and testing... see batch size issue annotations [#17]

+++ src/server/dori.rs
@@ -17,25 +17,39 @@ impl DoriServer {
 
         let service = AdaptiveSamplingService::new(config);
 
-        let uds_path_parent_dir = config.dori.uds_path.parent().unwrap();
+        if config.dori.tcp_enabled {
 
-        if config.dori.uds_path.exists() && config.dori.uds_path_override {
-            std::fs::remove_file(&config.dori.uds_path)?;
-            log::warn!("UDS override configured! Replaced existing socket: {}", config.dori.uds_path.display());
+            let address = format!("{}:{}", config.dori.tcp_host, config.dori.tcp_port).parse().unwrap();
+
+            log::info!("Dori TCP connection listening on: {}", address);
+
+            Server::builder()
+                .add_service(AdaptiveSamplingServer::new(service))
+                .serve(address)
+                .await?;
+            
+        } else {
+            let uds_path_parent_dir = config.dori.uds_path.parent().unwrap();
+
+            if config.dori.uds_path.exists() && config.dori.uds_path_override {
+                std::fs::remove_file(&config.dori.uds_path)?;
+                log::warn!("UDS override configured! Replaced existing socket: {}", config.dori.uds_path.display());
+            }
+    
+            if !uds_path_parent_dir.exists() {
+                std::fs::create_dir_all(config.dori.uds_path.parent().unwrap())?;
+                log::debug!("UDS parent directory created at: {}", &config.dori.uds_path.display());
+            }
+    
+            let uds = UnixListener::bind(&config.dori.uds_path)?;
+            let uds_stream = UnixListenerStream::new(uds);
+    
+            Server::builder()
+                .add_service(AdaptiveSamplingServer::new(service))
+                .serve_with_incoming(uds_stream)
+                .await?;
         }
-
-        if !uds_path_parent_dir.exists() {
-            std::fs::create_dir_all(config.dori.uds_path.parent().unwrap())?;
-            log::debug!("UDS parent directory created at: {}", &config.dori.uds_path.display());
-        }
-
-        let uds = UnixListener::bind(&config.dori.uds_path)?;
-        let uds_stream = UnixListenerStream::new(uds);
-
-        Server::builder()
-            .add_service(AdaptiveSamplingServer::new(service))
-            .serve_with_incoming(uds_stream)
-            .await?;
+        
 
         Ok(())
     }
@@ -47,15 +61,30 @@ pub struct DoriClient {
 
 impl DoriClient {
     pub async fn connect(config: &StreamfishConfig) -> Result<Self, Box<dyn std::error::Error>> {
-    
-        let uds_path = config.dori.uds_path.clone();
+        
+        if config.dori.tcp_enabled {
             
-        // We will ignore this URI because UDS do not use it
-        let channel = Endpoint::try_from("http://[::]:50051")?
-            .connect_with_connector(service_fn(move |_|  { 
-                UnixStream::connect(uds_path.clone()) 
-        })).await?;
+            let address = format!("http://{}:{}", config.dori.tcp_host, config.dori.tcp_port);
 
-        Ok(Self { client: AdaptiveSamplingClient::new(channel) })
+            log::info!("Dori client connecting to: {}", &address);
+
+            let channel = Channel::from_shared(address)?.connect().await?;
+
+            Ok(Self {  client: AdaptiveSamplingClient::new(channel) })
+
+        } else {
+            let uds_path = config.dori.uds_path.clone();
+            
+            // We will ignore this URI because UDS do not use it
+            let channel = Endpoint::try_from("http://[::]:50051")?
+                .connect_with_connector(service_fn(move |_|  { 
+                    UnixStream::connect(uds_path.clone()) 
+            })).await?;
+
+            Ok(Self { client: AdaptiveSamplingClient::new(channel) })
+        }
+        
+
+        
     }
 }
